backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
pymilvus==2.3.6
protobuf==4.25.2
grpcio==1.60.1
grpcio-tools==1.60.1
python-multipart==0.0.9
requests==2.31.0
numpy==1.26.4
pandas==2.2.0
dspy-ai==2.1.9
cryptography==42.0.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
prometheus-client==0.19.0
httpx==0.26.0
py_ecc==7.0.0
